Better handling of windows messages (and more compatible with Delphi....)
procedure WMCOmmand(); message WM_COMMAND;
is done in this way:
[MessageAttribute(rtl.WM_COMMAND)]
method WMCommand(var aMessage: TMessage);

each class that has methods of this kind has a class constructor to build the message method table. Then, when processing the message, we can walk the current type table and find if there is a method registered for the current message. If so, then call it!